partial support for tracepoints in standard queries

+++ src/partition.rs
@@ -25,7 +25,7 @@ pub fn partition_alignments(
     min_distance_between_ranges: i32,
     output_format: &str,
     fasta_index: Option<&FastaIndex>,
-    scoring_params: Option<(u8, u8, u8, u8, u8, u8)>,
+    scoring_params: Option<(i32, i32, i32, i32, i32, i32)>,
     debug: bool,
 ) -> io::Result<()> {
     // Initialize windows from starting sequences if provided
@@ -1074,7 +1074,7 @@ fn write_partition(
     impg: &Impg,
     output_format: &str,
     fasta_index: Option<&FastaIndex>,
-    scoring_params: Option<(u8, u8, u8, u8, u8, u8)>,
+    scoring_params: Option<(i32, i32, i32, i32, i32, i32)>,
 ) -> io::Result<()> {
     match output_format {
         "bed" => write_partition_bed(partition_num, query_intervals, impg, None),
@@ -1163,7 +1163,7 @@ fn write_partition_gfa(
     query_intervals: &[Interval<u32>],
     impg: &Impg,
     fasta_index: &FastaIndex,
-    scoring_params: (u8, u8, u8, u8, u8, u8),
+    scoring_params: (i32, i32, i32, i32, i32, i32),
 ) -> io::Result<()> {
     // Generate a GFA-formatted string from the list of intervals
     let gfa_output = crate::graph::generate_gfa_from_intervals(
@@ -1188,7 +1188,7 @@ fn write_partition_maf(
     query_intervals: &[Interval<u32>],
     impg: &Impg,
     fasta_index: &FastaIndex,
-    scoring_params: (u8, u8, u8, u8, u8, u8),
+    scoring_params: (i32, i32, i32, i32, i32, i32),
 ) -> io::Result<()> {
     // Generate a MAF-formatted string from the list of intervals
     let maf_output = crate::graph::generate_maf_from_intervals(
